requirements.txt
fastapi
uvicorn[standard]
striprtf
python-dotenv
